requirements.txt
requests
pytubefix>=2.4.0
azure-identity>=1.12.0
ffmpeg
ffmpeg-python
moviepy
azure-storage-blob>=12.17.0
python-dotenv>=1.0.0
firebase-admin>=6.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0
pytube>=12.0.0
openai>=1.1.0
replicate>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
